+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - [refactor:optimize auto configuration.](https://github.com/Tencent/spring-cloud-tencent/pull/1740)
 - [refactor:optimize config locate.](https://github.com/Tencent/spring-cloud-tencent/pull/1742)
 - [feat:support async metadata transfer.](https://github.com/Tencent/spring-cloud-tencent/pull/1744)
+- [feat: support TagUtils, ContextToHeaderInterceptor in TSF.](https://github.com/Tencent/spring-cloud-tencent/pull/1753)